+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 - Fix some minor issues in landing (PATCH)
 
+# 1.3.0
+
+- Crypto value error issue on send modal resolved (MINOR)
+- Add user through partner role (PATCH)
+- Fetch roles isues by admin and partner role (PATCH)
+- Wallet drawer close issue in send modal resolved (MINOR)
+- Disconnect wallet on convertor modal closes (MINOR)
+- update step 2 values in send modal (PATCH)
+
 # v1.2.0
 
 - Settings Page is implemented (MAJOR)